api/pyproject.toml
[tool.poetry]
name = "zh-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "0.19.0"
fastapi = "0.87.0"
psycopg2-binary = "2.9.5"
SQLAlchemy = "1.4.41"
